pyproject.toml
[project]
name = "ai-art-generator"
version = "0.1.0"
description = "A Python CLI tool for generating visual assets using multiple AI image generation services"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.20.0",
    "google-cloud-aiplatform>=1.73.0",
    "openai>=1.88.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
